requirements.txt
amqp                          5.1.1
asgiref                       3.5.2
billiard                      3.6.4.0
celery                        5.2.7
click                         8.1.3
click-didyoumean              0.3.0
click-plugins                 1.1.1
click-repl                    0.2.0
Django                        4.0.8
django-celery-beat            2.3.0
django-celery-results         2.4.0
django-timezone-field         5.0
djangorestframework           3.14.0
djangorestframework-simplejwt 5.2.1
kombu                         5.2.4
pip                           22.0.4
prompt-toolkit                3.0.31
PyJWT                         2.5.0
python-crontab                2.6.0
python-dateutil               2.8.2
pytz                          2022.4
setuptools                    58.1.0
six                           1.16.0
sqlparse                      0.4.3
tzdata                        2022.4
vine                          5.0.0
wcwidth                       0.2.5